Lift parsing of results out of executor implementations
