+++ trac-libs/trac-lib-exec/src/main/java/com/accenture/trac/common/exec/IBatchExecutor.java
@@ -37,7 +37,7 @@ public interface IBatchExecutor {
 
     void getBatchStatus(String jobKey, ExecutorState jobState);
 
-    ExecutorPollResult readBatchResult(String jobKey, ExecutorState jobState);
+    byte[] readBatchResult(String jobKey, ExecutorState jobState);
 
     ExecutorState cancelBatch(String jobKey, ExecutorState jobState);
