+++ trac-libs/trac-lib-exec/src/main/java/com/accenture/trac/common/exec/kubernetes/KubernetesBatchExecutor.java
@@ -356,7 +356,7 @@ public class KubernetesBatchExecutor implements IBatchExecutor {
     }
 
     @Override
-    public ExecutorPollResult readBatchResult(String jobKey, ExecutorState jobState) {
+    public byte[] readBatchResult(String jobKey, ExecutorState jobState) {
 
 
         // kubeBatchApi.
